+++ egui/src/app.rs
@@ -55,7 +55,7 @@ pub struct IntegrationContext<'a> {
     pub tex_allocator: Option<&'a mut dyn TextureAllocator>,
     /// Where the app can issue commands back to the integration.
     pub output: AppOutput,
-    /// If you need to request a repaint from another thread, clone this and give to that other thread
+    /// If you need to request a repaint from another thread, clone this and send it to that other thread.
     pub repaint_signal: std::sync::Arc<dyn RepaintSignal>,
 }
 
@@ -113,7 +113,7 @@ pub trait TextureAllocator {
     fn free(&mut self, id: crate::TextureId);
 }
 
-pub trait RepaintSignal: Send {
+pub trait RepaintSignal: Send + Sync {
     /// This signals the Egui integration that a repaint is required.
     /// This is meant to be called when a background process finishes in an async context and/or background thread.
     fn request_repaint(&self);