+++ CHANGELOG.md
@@ -9,6 +9,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Fixed 🐛
+
+* `RepaintSignal` now implements `Sync` so it can be sent to a background thread.
+
+
 ## 0.6.0 - 2020-12-26
 
 ### Added ⭐